refactor(ios): VoiceAlarmAPI/LocalAlarmStore/VoiceProfileManagementPanel 분할
- VoiceAlarmAPI(1864) → +VoiceAlarmAPIModels.swift (모델 구조체 분리)
- LocalAlarmStore(936) → +LocalAlarmRecord.swift (레코드/검증 enum 분리)
- VoiceProfileManagementPanel(886) → +VoiceProfileManagementComponents.swift
  (행/다이얼로그 하위 컴포넌트 분리, private→internal)

동작/디자인 변경 없음 — 동일 모듈 내 파일/가시성만 조정.
