+++ apps/ios-native/VoiceAlarm/Views/Voices/VoiceProfileManagementComponents.swift
@@ -1,0 +1,578 @@
+import SwiftUI
+
+// VoiceProfileManagementPanel 에서 분리한 행/다이얼로그 하위 컴포넌트.
+// 동작/디자인 변경 없음 — internal 가시성만 조정.
+
+// MARK: - Row
+
+/// 단일 목소리 프로필 행. Android `VoiceProfileRow` 와 동일한 슬롯.
+struct VoiceProfileRow: View {
+    let profile: VoiceProfile
+    let isSelected: Bool
+    let canShareVoice: Bool
+    let onSelect: () -> Void
+    let onEdit: () -> Void
+    let onDelete: () -> Void
+    let onToggleShare: (Bool) -> Void
+
+    var body: some View {
+        VStack(alignment: .leading, spacing: 10) {
+            HStack(spacing: 12) {
+                ZStack {
+                    Circle().fill(VoiceAlarmTheme.surfaceVariant)
+                    Image(systemName: "mic")
+                        .foregroundStyle(VoiceAlarmTheme.primary)
+                }
+                .frame(width: 40, height: 40)
+
+                VStack(alignment: .leading, spacing: 3) {
+                    Text(profile.name)
+                        .font(.subheadline.weight(.semibold))
+                        .foregroundStyle(VoiceAlarmTheme.text)
+                    HStack(spacing: 6) {
+                        if shouldShowStatusPill {
+                            statusPill
+                        }
+                        if profile.isShared == true {
+                            sharedPill
+                        }
+                    }
+                }
+                Spacer()
+                Menu {
+                    Button("선택", action: onSelect)
+                    Button("정보 수정", action: onEdit)
+                    if canShareVoice {
+                        Toggle("공유 허용", isOn: Binding(
+                            get: { profile.isShared ?? false },
+                            set: { onToggleShare($0) }
+                        ))
+                    } else {
+                        Button("공유 허용") {}
+                            .disabled(true)
+                    }
+                    Divider()
+                    Button("삭제", role: .destructive, action: onDelete)
+                } label: {
+                    Image(systemName: "ellipsis.circle")
+                        .font(.title3)
+                        .foregroundStyle(VoiceAlarmTheme.textSecondary)
+                        .frame(width: 36, height: 36)
+                }
+            }
+
+            HStack(spacing: 8) {
+                Button(isSelected ? "선택됨" : "이 목소리 사용", action: onSelect)
+                    .buttonStyle(.bordered)
+                    .tint(isSelected ? VoiceAlarmTheme.primary : VoiceAlarmTheme.textSecondary)
+                    .disabled(isSelected)
+                Spacer()
+                Button {
+                    onEdit()
+                } label: {
+                    Image(systemName: "pencil")
+                }
+                .buttonStyle(.borderless)
+                .foregroundStyle(VoiceAlarmTheme.primary)
+                Button(role: .destructive) {
+                    onDelete()
+                } label: {
+                    Image(systemName: "trash")
+                }
+                .buttonStyle(.borderless)
+                .foregroundStyle(VoiceAlarmTheme.error)
+            }
+        }
+        .padding(14)
+        .background(VoiceAlarmTheme.surface)
+        .overlay(
+            RoundedRectangle(cornerRadius: 12)
+                .stroke(isSelected ? VoiceAlarmTheme.primary : VoiceAlarmTheme.outline, lineWidth: isSelected ? 1.5 : 1)
+        )
+        .clipShape(RoundedRectangle(cornerRadius: 12))
+    }
+
+    private var shouldShowStatusPill: Bool {
+        let status = normalizedStatus(profile.status)
+        return status != "ready"
+    }
+
+    private var statusPill: some View {
+        let status = normalizedStatus(profile.status)
+        let bg: Color
+        let fg: Color
+        switch status {
+        case "ready":
+            bg = VoiceAlarmTheme.accent.opacity(0.18); fg = VoiceAlarmTheme.accent
+        case "processing":
+            bg = VoiceAlarmTheme.secondary.opacity(0.18); fg = VoiceAlarmTheme.secondary
+        case "deleting":
+            bg = VoiceAlarmTheme.surfaceVariant; fg = VoiceAlarmTheme.textSecondary
+        case "failed":
+            bg = VoiceAlarmTheme.error.opacity(0.18); fg = VoiceAlarmTheme.error
+        default:
+            bg = VoiceAlarmTheme.surfaceVariant; fg = VoiceAlarmTheme.textSecondary
+        }
+        return Text(statusLabel(status))
+            .font(.caption.weight(.semibold))
+            .padding(.horizontal, 8)
+            .padding(.vertical, 3)
+            .background(bg, in: Capsule())
+            .foregroundStyle(fg)
+    }
+
+    private var sharedPill: some View {
+        Text("공유 중")
+            .font(.caption.weight(.semibold))
+            .padding(.horizontal, 8)
+            .padding(.vertical, 3)
+            .background(VoiceAlarmTheme.primary.opacity(0.15), in: Capsule())
+            .foregroundStyle(VoiceAlarmTheme.primary)
+    }
+
+    private func statusLabel(_ raw: String) -> String {
+        switch raw {
+        case "ready": return "사용 가능"
+        case "processing": return "학습 중"
+        case "deleting": return "삭제 중"
+        case "failed": return "실패"
+        default: return "상태 확인 중"
+        }
+    }
+
+    private func normalizedStatus(_ raw: String?) -> String {
+        let status = raw?.trimmingCharacters(in: .whitespacesAndNewlines) ?? ""
+        return status.isEmpty ? "ready" : status
+    }
+}
+
+struct FamilyVoiceProfileRow: View {
+    let profile: FamilyVoiceProfile
+    let onEdit: () -> Void
+
+    /// Android `SharedVoiceProfileRow.needsViewerInfo` (VoiceProfileManagementPanel.kt:1477) 와 동일.
+    /// 관계/호칭 중 하나라도 비어 있으면 "설정하기" CTA 버튼을 노출.
+    private var needsViewerInfo: Bool {
+        profile.requiresViewerInfo
+    }
+
+    var body: some View {
+        VStack(alignment: .leading, spacing: 10) {
+            HStack(spacing: 12) {
+                ZStack {
+                    Circle().fill(VoiceAlarmTheme.secondary.opacity(0.18))
+                    Image(systemName: "person.wave.2")
+                        .foregroundStyle(VoiceAlarmTheme.secondary)
+                }
+                .frame(width: 40, height: 40)
+                VStack(alignment: .leading, spacing: 3) {
+                    Text(profile.name).font(.subheadline.weight(.semibold))
+                    Text(profile.sharedFromLabel)
+                        .font(.caption)
+                        .foregroundStyle(VoiceAlarmTheme.textSecondary)
+                }
+                Spacer()
+                Button(action: onEdit) {
+                    Image(systemName: "pencil")
+                        .font(.callout)
+                        .foregroundStyle(VoiceAlarmTheme.primary)
+                        .frame(width: 36, height: 36)
+                }
+                .buttonStyle(.borderless)
+                .accessibilityLabel("내 정보 수정")
+            }
+            if needsViewerInfo {
+                Button(action: onEdit) {
+                    Text("이 목소리가 나를 어떻게 부를지 설정")
+                        .font(.footnote.weight(.semibold))
+                        .frame(maxWidth: .infinity)
+                }
+                .buttonStyle(.bordered)
+                .tint(VoiceAlarmTheme.primary)
+            }
+        }
+        .padding(12)
+        .background(VoiceAlarmTheme.surface)
+        .overlay(
+            RoundedRectangle(cornerRadius: 12).stroke(VoiceAlarmTheme.outline, lineWidth: 1)
+        )
+        .clipShape(RoundedRectangle(cornerRadius: 12))
+    }
+
+}
+
+// MARK: - Edit dialog
+
+/// 프로필 이름 + 공유 토글 동시 편집 다이얼로그.
+struct VoiceProfileEditDialog: View {
+    let initialName: String
+    let initialRelationship: String
+    let initialListenerTitle: String
+    let initialIsShared: Bool
+    let canShareVoice: Bool
+    let onCancel: () -> Void
+    let onSave: (String, String, String, Bool) -> Void
+
+    @State private var name: String = ""
+    @State private var relationshipSelection = VoiceRelationshipSelection()
+    @State private var listenerTitle: String = ""
+    @State private var isShared: Bool = false
+    @State private var submitted = false
+
+    private var trimmedName: String {
+        name.trimmingCharacters(in: .whitespacesAndNewlines)
+    }
+
+    private var trimmedRelationship: String {
+        relationshipSelection.resolved
+    }
+
+    private var trimmedListenerTitle: String {
+        listenerTitle.trimmingCharacters(in: .whitespacesAndNewlines)
+    }
+
+    var body: some View {
+        VStack(alignment: .leading, spacing: 16) {
+            HStack(alignment: .top) {
+                Text("정보 수정")
+                    .font(.title3.weight(.bold))
+                Spacer()
+                Button(action: onCancel) {
+                    Image(systemName: "xmark")
+                        .foregroundStyle(VoiceAlarmTheme.textSecondary)
+                }
+                .buttonStyle(.plain)
+                .accessibilityLabel(Text("닫기"))
+            }
+
+            VStack(alignment: .leading, spacing: 6) {
+                Text("이름")
+                    .font(.caption)
+                    .foregroundStyle(VoiceAlarmTheme.textSecondary)
+                TextField("목소리 이름", text: $name)
+                    .textFieldStyle(.roundedBorder)
+                    .onChange(of: name) { _, newValue in
+                        if newValue.count > 50 {
+                            name = String(newValue.prefix(50))
+                        }
+                    }
+                if submitted && trimmedName.isEmpty {
+                    Text("목소리 이름을 입력해 주세요.")
+                        .font(.caption2.weight(.semibold))
+                        .foregroundStyle(VoiceAlarmTheme.error)
+                }
+            }
+
+            VoiceRelationshipInputField(
+                selection: $relationshipSelection,
+                submitted: submitted
+            )
+
+            VStack(alignment: .leading, spacing: 6) {
+                Text("이 목소리가 나를 부를 호칭")
+                    .font(.caption)
+                    .foregroundStyle(VoiceAlarmTheme.textSecondary)
+                TextField("예: 지호야, 여보", text: $listenerTitle)
+                    .textFieldStyle(.roundedBorder)
+                    .onChange(of: listenerTitle) { _, newValue in
+                        if newValue.count > 30 {
+                            listenerTitle = String(newValue.prefix(30))
+                        }
+                    }
+                if submitted && trimmedListenerTitle.isEmpty {
+                    Text("이 목소리가 나를 부를 이름을 입력해 주세요.")
+                        .font(.caption2.weight(.semibold))
+                        .foregroundStyle(VoiceAlarmTheme.error)
+                }
+            }
+
+            VoiceListenerPreviewCard(
+                listenerTitle: listenerTitle,
+                relationshipLabel: trimmedRelationship
+            )
+
+            Toggle(isOn: $isShared) {
+                VStack(alignment: .leading, spacing: 2) {
+                    Text("목소리 공유").font(.subheadline.weight(.semibold))
+                    Text(shareDescription)
+                        .font(.caption)
+                        .foregroundStyle(VoiceAlarmTheme.textSecondary)
+                }
+            }
+            .disabled(!canShareVoice)
+
+            Button("저장") {
+                submitted = true
+                guard !trimmedName.isEmpty,
+                      !trimmedRelationship.isEmpty,
+                      !trimmedListenerTitle.isEmpty else {
+                    return
+                }
+                onSave(trimmedName, trimmedRelationship, trimmedListenerTitle, isShared)
+            }
+            .buttonStyle(.borderedProminent)
+            .tint(VoiceAlarmTheme.primary)
+            .frame(maxWidth: .infinity)
+            Spacer(minLength: 0)
+        }
+        .padding(20)
+        .onAppear {
+            name = initialName
+            relationshipSelection = parseVoiceRelationshipLabel(initialRelationship)
+            listenerTitle = initialListenerTitle
+            isShared = initialIsShared
+        }
+    }
+
+    private var shareDescription: String {
+        if !canShareVoice {
+            return "공유는 커플/가족 이용권에서 사용할 수 있어요."
+        }
+        return isShared ? "이용권을 같이 사용하는 사람들에게 목소리를 공유해요." : "내 계정에서만 사용해요."
+    }
+}
+
+// MARK: - Delete dialog
+
+/// 삭제 확인 다이얼로그. force 토글 + 영향받는 알람 수 안내.
+struct VoiceProfileDeleteDialog: View {
+    let profileName: String
+    @Binding var force: Bool
+    let onCancel: () -> Void
+    let onConfirm: () -> Void
+
+    var body: some View {
+        VStack(alignment: .leading, spacing: 16) {
+            HStack(alignment: .top) {
+                Text("목소리 삭제")
+                    .font(.title3.weight(.bold))
+                Spacer()
+                Button(action: onCancel) {
+                    Image(systemName: "xmark")
+                        .foregroundStyle(VoiceAlarmTheme.textSecondary)
+                }
+            }
+
+            Text("'\(profileName)' 목소리를 삭제할까요?")
+                .font(.subheadline)
+            Text("이 목소리를 쓰는 메시지는 텍스트만 남고, 알람은 기본 알람음으로 바뀌어요. 저장된 음원 파일도 함께 삭제돼요.")
+                .font(.footnote)
+                .foregroundStyle(VoiceAlarmTheme.textSecondary)
+
+            Toggle(isOn: $force) {
+                VStack(alignment: .leading, spacing: 2) {
+                    Text("사용 중인 알람도 함께 정리").font(.subheadline.weight(.semibold))
+                    Text(force ? "기본 알람음으로 강등돼요" : "사용 중이면 삭제하지 않아요")
+                        .font(.caption)
+                        .foregroundStyle(VoiceAlarmTheme.textSecondary)
+                }
+            }
+
+            HStack {
+                Button("취소", action: onCancel)
+                    .buttonStyle(.bordered)
+                    .frame(maxWidth: .infinity)
+                Button(role: .destructive, action: onConfirm) {
+                    Text("삭제")
+                        .frame(maxWidth: .infinity)
+                }
+                .buttonStyle(.borderedProminent)
+                .tint(VoiceAlarmTheme.error)
+            }
+            Spacer(minLength: 0)
+        }
+        .padding(20)
+    }
+}
+
+// MARK: - Shared voice viewer info dialog
+
+/// 공유받은 음성에 대해 viewer 가 자신과의 관계와 호칭을 설정하는 다이얼로그.
+///
+/// Android `SharedVoiceViewerInfoDialog`
+/// (`VoiceProfileManagementPanel.kt:1543`) 와 1:1 대응. 둘 다 필수이며, 비어 있으면
+/// 인라인 오류 메시지를 띄우고 저장을 막는다.
+struct SharedVoiceViewerInfoDialog: View {
+    let profileName: String
+    let sharedFromLabel: String
+    let initialRelationship: String
+    let initialListenerTitle: String
+    let isWorking: Bool
+    let onCancel: () -> Void
+    let onPreview: () -> Void
+    let onConfirm: (String, String) -> Void
+
+    @State private var relationshipSelection = VoiceRelationshipSelection()
+    @State private var listenerTitle: String = ""
+    @State private var submitted: Bool = false
+
+    private var trimmedRelationship: String {
+        relationshipSelection.resolved
+    }
+    private var trimmedListener: String {
+        listenerTitle.trimmingCharacters(in: .whitespacesAndNewlines)
+    }
+    private var listenerError: Bool { submitted && trimmedListener.isEmpty }
+
+    var body: some View {
+        VStack(alignment: .leading, spacing: 16) {
+            HStack(alignment: .top) {
+                Text("공유받은 목소리 설정")
+                    .font(.title3.weight(.bold))
+                Spacer()
+                Button(action: onCancel) {
+                    Image(systemName: "xmark")
+                        .foregroundStyle(VoiceAlarmTheme.textSecondary)
+                }
+                .buttonStyle(.plain)
+            }
+            Text("'\(profileName)' 가 내게 어떻게 말할지 알려주세요.")
+                .font(.subheadline)
+                .foregroundStyle(VoiceAlarmTheme.textSecondary)
+
+            HStack(spacing: 12) {
+                ZStack {
+                    Circle().fill(VoiceAlarmTheme.secondary.opacity(0.18))
+                    Image(systemName: "mic")
+                        .foregroundStyle(VoiceAlarmTheme.secondary)
+                }
+                .frame(width: 44, height: 44)
+                VStack(alignment: .leading, spacing: 3) {
+                    Text(profileName)
+                        .font(.headline)
+                    Text(sharedFromLabel)
+                        .font(.caption)
+                        .foregroundStyle(VoiceAlarmTheme.textSecondary)
+                }
+                Spacer(minLength: 0)
+            }
+            .padding(14)
+            .background(VoiceAlarmTheme.surfaceVariant.opacity(0.55))
+            .overlay(
+                RoundedRectangle(cornerRadius: 16).stroke(VoiceAlarmTheme.outline, lineWidth: 1)
+            )
+            .clipShape(RoundedRectangle(cornerRadius: 16))
+
+            VoiceRelationshipInputField(
+                selection: $relationshipSelection,
+                submitted: submitted
+            )
+
+            VStack(alignment: .leading, spacing: 6) {
+                Text("이 목소리가 나를 부를 호칭")
+                    .font(.caption.weight(.semibold))
+                    .foregroundStyle(VoiceAlarmTheme.textSecondary)
+                TextField("예: 지호야, 우리 강아지", text: $listenerTitle)
+                    .textFieldStyle(.roundedBorder)
+                    .onChange(of: listenerTitle) { _, newValue in
+                        if newValue.count > 30 {
+                            listenerTitle = String(newValue.prefix(30))
+                        }
+                    }
+                if listenerError {
+                    Text("필수 입력 값입니다.")
+                        .font(.caption2.weight(.semibold))
+                        .foregroundStyle(VoiceAlarmTheme.error)
+                }
+            }
+
+            VoiceListenerPreviewCard(
+                listenerTitle: listenerTitle,
+                relationshipLabel: trimmedRelationship
+            )
+
+            VStack(spacing: 8) {
+                Button(action: onPreview) {
+                    Label("미리듣기", systemImage: "play.fill")
+                        .frame(maxWidth: .infinity)
+                }
+                .buttonStyle(.bordered)
+                .disabled(isWorking)
+
+                Button("저장") {
+                    submitted = true
+                    if !trimmedRelationship.isEmpty && !trimmedListener.isEmpty {
+                        onConfirm(trimmedRelationship, trimmedListener)
+                    }
+                }
+                .buttonStyle(.borderedProminent)
+                .tint(VoiceAlarmTheme.primary)
+                .frame(maxWidth: .infinity)
+                .disabled(isWorking)
+            }
+            Spacer(minLength: 0)
+        }
+        .padding(20)
+        .onAppear {
+            relationshipSelection = parseVoiceRelationshipLabel(initialRelationship)
+            listenerTitle = initialListenerTitle
+        }
+    }
+}
+
+// MARK: - Plan gate
+
+/// 슬롯 부족 / 유료 플랜 필요 시 노출하는 안내 시트.
+struct VoicePlanGateSheet: View {
+    let onUpgrade: () -> Void
+    let onClose: () -> Void
+
+    var body: some View {
+        VStack(alignment: .leading, spacing: 16) {
+            HStack {
+                Text("플랜 업그레이드가 필요해요")
+                    .font(.title3.weight(.bold))
+                Spacer()
+                Button(action: onClose) {
+                    Image(systemName: "xmark")
+                        .foregroundStyle(VoiceAlarmTheme.textSecondary)
+                }
+                .accessibilityLabel(Text("닫기"))
+            }
+            Text("목소리 슬롯이 가득 찼거나, 본 기능은 유료 플랜에서 사용할 수 있어요.")
+                .font(.subheadline)
+                .foregroundStyle(VoiceAlarmTheme.textSecondary)
+            VStack(alignment: .leading, spacing: 8) {
+                bullet("기존 목소리를 삭제해 자리를 만들어요")
+                bullet("Family · Couple 플랜으로 업그레이드해 슬롯을 확장해요")
+            }
+            Button("플랜 보기", action: onUpgrade)
+                .buttonStyle(.borderedProminent)
+                .tint(VoiceAlarmTheme.primary)
+                .frame(maxWidth: .infinity)
+            Spacer(minLength: 0)
+        }
+        .padding(20)
+    }
+
+    private func bullet(_ text: String) -> some View {
+        HStack(alignment: .top, spacing: 6) {
+            Image(systemName: "checkmark.circle.fill")
+                .foregroundStyle(VoiceAlarmTheme.primary)
+            Text(text).font(.footnote)
+        }
+    }
+}
+
+#if DEBUG
+#Preview("VoiceProfileManagementPanel (light)") {
+    VoiceProfileManagementPanel(route: .constant(.management))
+        .voiceAlarmPreviewEnvironment()
+}
+
+#Preview("VoiceProfileManagementPanel (dark)") {
+    VoiceProfileManagementPanel(route: .constant(.management))
+        .preferredColorScheme(.dark)
+        .voiceAlarmPreviewEnvironment()
+}
+
+#Preview("Delete dialog") {
+    VoiceProfileDeleteDialog(
+        profileName: "아침 목소리",
+        force: .constant(true),
+        onCancel: {},
+        onConfirm: {}
+    )
+    .padding()
+}
+#endif
